fix: HTTPS NIC tests - TLS encryption, throughput architecture, and test assertions
- Fix TLS data path: encrypt HTTP requests after handshake, decrypt
  responses before parsing, add app_state 5 to decrypt condition
- Rewrite throughput to use IPC (worker-side execution) instead of
  mgmt lcore, preventing TX queue race condition
- Add dedicated mgmt TX queue (n+1 TX queues in port_init)
- Fix throughput marker segfault (app_ctx guard for sentinel value)
- Reduce TLS buffer to 1400 bytes to fit in MSS after encryption
- Batch RSTs in cmd_reset with 5ms delays to prevent TX ring overflow
- Switch throughput discard server to socat OPENSSL on port 5001
- Fix test assertions: 5xx under flood is warn, TARGET_CPS=400,
  remove http_req_tx check from throughput tests
- Fix build: sign-compare cast in tcp_options.c, missing include in
  worker_loop.c

All 17 HTTPS NIC tests pass with QAT offload on both sides.

+++ vaigAI/src/net/tcp_options.c
@@ -130,7 +130,7 @@ int tcp_options_write_data(uint8_t *buf, size_t bufsz,
     uint8_t *p = buf;
     size_t   remaining = bufsz;
 
-#define NEED(n) if (remaining < (n)) goto done; remaining -= (n);
+#define NEED(n) if (remaining < (size_t)(n)) goto done; remaining -= (size_t)(n);
 
     if (timestamps) {
         NEED(12);
